requirements.txt
boto3>=1.28.0
botocore>=1.31.0
pyyaml>=6.0
jinja2>=3.1.2
requests>=2.31.0